requirements.txt
Django==3.0.8
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
django-model-utils==4.0.0
psycopg2==2.9.2
jsbeautifier==1.14.0